Added point cloud streaming to Kinect 2 Client
Added functionality to the Kinect2ServerService frame recieved callback function to convert
a depth image frame to a point cloud and broadcast it as a byte array over a new port, as
defined in the app.config. This commit should be viewed in parallel with changes to
the K2 client package, which adds a new node allowing the client to recieve and deserialise
point cloud data.